+++ driver-core/CHANGELOG.rst
@@ -8,6 +8,7 @@ CHANGELOG
 - [improvement] Defunct connection on request timeout (JAVA-425)
 - [improvement] Try next host when we get a SERVER_ERROR (JAVA-426)
 - [bug] Handle race between query timeout and completion (JAVA-449, JAVA-460, JAVA-471)
+- [bug] Fix DCAwareRoundRobinPolicy datacenter auto-discovery (JAVA-496)
 
 
 2.0.6: